==> analysis_merfish/71.check_data_250408_which_is_which.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import anndata as ad
import scanpy as sc
from scipy import stats
import os

from sklearn.decomposition import PCA
from sklearn.cluster import KMeans
from umap import UMAP

from scroutines import powerplots
from scroutines.miscu import is_in_polygon

# load input

from merfish_datasets import merfish_datasets as directories

outdir     = "/u/home/f/f7xiesnm/project-zipursky/v1-bb/v1/results_merfish/plots_250408"
outdatadir = "/u/home/f/f7xiesnm/project-zipursky/v1-bb/v1/data/merfish/organized"
!mkdir -p $outdir
!mkdir -p $outdatadir

dfs = []
# merfish_202503111431
names = ['P21NRc_ant', 'P8NRa_ant', 'P8NRb_ant', 'P8NRc_ant', 'P8NRd_ant',]
for name in names:
    dirc = directories[name]
    f2 = os.path.join('/u/home/f/f7xiesnm/project-zipursky/v1-bb/v1/data/merfish', dirc, 'cell_metadata.csv')
    df2 = pd.read_csv(f2)
    dfs.append(df2)
    
fig, ax = plt.subplots()
for i in range(len(dfs)): 
    x = dfs[i]['center_x']
    y = dfs[i]['center_y']
    ax.scatter(x, y, s=0.05, edgecolor='none')
ax.invert_yaxis()
plt.show()

dfs = []
# merfish_202503111242
names = ['P21NRc_pos', 'P8NRa_pos', 'P8NRb_pos', 'P8NRc_pos', 'R5_pos',]
for name in names:
    dirc = directories[name]
    f2 = os.path.join('/u/home/f/f7xiesnm/project-zipursky/v1-bb/v1/data/merfish', dirc, 'cell_metadata.csv')
    df2 = pd.read_csv(f2)
    dfs.append(df2)
    
fig, ax = plt.subplots()
for i in range(len(dfs)): 
    x = dfs[i]['center_x']
    y = dfs[i]['center_y']
    ax.scatter(x, y, s=0.05, edgecolor='none')
ax.invert_yaxis()
plt.show()

dfs = []
# exp merfish_202503191505 (repeated experiments) same animals diff. sections? 
names = ['P21NRc_ant2', 'P8NRa_ant2', 'P8NRb_ant2', 'P8NRc_ant2', 'P8NRd_ant2',]
for name in names:
    dirc = directories[name]
    f2 = os.path.join('/u/home/f/f7xiesnm/project-zipursky/v1-bb/v1/data/merfish', dirc, 'cell_metadata.csv')
    df2 = pd.read_csv(f2)
    dfs.append(df2)
    
fig, ax = plt.subplots()
for i in range(len(dfs)): 
    x = dfs[i]['center_x']
    y = dfs[i]['center_y']
    ax.scatter(x, y, s=0.05, edgecolor='none')
ax.invert_yaxis()
plt.show()

dfs = []
# exp merfish_202503191540
names = ['P21NRc_pos2', 'P21NRc_pos2a', 'P8NRa_pos2', 'P8NRb_pos2', 'P8NRc_pos2', 'P8NRd_pos2',]
for name in names:
    dirc = directories[name]
    f2 = os.path.join('/u/home/f/f7xiesnm/project-zipursky/v1-bb/v1/data/merfish', dirc, 'cell_metadata.csv')
    df2 = pd.read_csv(f2)
    dfs.append(df2)
    
fig, ax = plt.subplots()
for i in range(len(dfs)): 
    x = dfs[i]['center_x']
    y = dfs[i]['center_y']
    ax.scatter(x, y, s=0.05, edgecolor='none')
ax.invert_yaxis()
plt.show()
